% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{download_tol}
\alias{download_tol}
\title{Download a file from the ToL SFTP server.}
\usage{
download_tol(download_link, save_dir = NULL)
}
\arguments{
\item{download_link}{A string specifying the URL to download the file from.
You can get a download URL for a particular specimen or gene using \code{\link[=lookup_tol]{lookup_tol()}}.}

\item{save_dir}{A string specifying the folder to save the download in. If
no value is provided, \link[here]{here} will be used.}
}
\description{
Download an alignment, sequence, or tree file from the ToL
SFTP server.
}
\details{
The \href{https://treeoflife.kew.org/}{Tree of Life} is a database
of specimens sequenced as part of Kew's efforts to build
a comprehensive evolutionary tree of life for flowering plants.

Sequence, alignment, and Newick tree files are help on an SFTP server
for download. The URLs to access these are stored in entries for specimens
and genes in the ToL database. These can be accessed by either using \code{\link[=search_tol]{search_tol()}}
to get all specimens for a particular order, family, genus, or species or by
looking up a specific specimen or gene using \code{\link[=lookup_tol]{lookup_tol()}}
}
\examples{
\dontrun{
 # download a specimen fasta file
 specimen_info <- lookup_tol("1296")
 download_tol(specimen_info$fasta_file_url)

 # download a gene alignment file
 gene_info <- lookup_tol("51", type="gene")
 download_tol(gene_info$alignment_file_url)

 # download the gene tree
 download_tol(gene_info$tree_file_url)
}

}
\references{
Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D., Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L., Kersey P.J., Leitch I.J. & Forest F. 2021. A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life. Systematic Biology, 2021; syab035, https://doi.org/10.1093/sysbio/syab035
}
\seealso{
\itemize{
\item \code{\link[=lookup_tol]{lookup_tol()}} to lookup information about a sequenced specimen
using a valid ToL ID.
\item \code{\link[=search_tol]{search_tol()}} to search ToL using taxonomic info.
}

Other ToL functions: 
\code{\link{lookup_tol}()}
}
\concept{ToL functions}
