% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{download_wcvp}
\alias{download_wcvp}
\title{Download the whole of the WCVP.}
\usage{
download_wcvp(save_dir = NULL, version = NULL)
}
\arguments{
\item{save_dir}{A string specifying the folder to save the download in. If
no value is provided, \link[here]{here} will be used.}

\item{version}{An integer version number to download. The latest
version will be downloaded by default.}
}
\description{
Download the latest or a specific version of the World
Checklist of Vascular Plants (WCVP).
}
\details{
The \href{https://wcvp.science.kew.org/}{World Checklist of Vascular Plants (WCVP)}
is a global consensus view of all known vascular plant species.
It has been compiled by staff at RBG Kew in consultation with plant
group experts.

Versioned downloads of the whole WCVP are provided on the website.
This function allows the user to download the latest or a specific
version of the WCVP.
}
\examples{
\dontrun{
 # download the latest version
 download_wcvp()

 # download version 1
 download_wcvp(version=1)
}

}
\references{
WCVP (2020). World Checklist of Vascular Plants, version 2.0. Facilitated by the Royal Botanic Gardens, Kew. Published on the Internet; http://wcvp.science.kew.org/
}
\seealso{
\itemize{
\item \code{\link[=lookup_wcvp]{lookup_wcvp()}} to lookup information about a taxon name
using a valid IPNI ID.
\item \code{\link[=search_wcvp]{search_wcvp()}} to search WCVP using a taxon name.
}

Other WCVP functions: 
\code{\link{lookup_wcvp}()},
\code{\link{search_wcvp}()}
}
\concept{WCVP functions}
