% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filters}
\alias{get_filters}
\title{Get valid filters for a resource.}
\usage{
get_filters(resource = c("wcvp", "powo", "ipni"))
}
\arguments{
\item{resource}{The resource being queried.}
}
\value{
A character vector of valid filters.
}
\description{
Get valid filters for a resource.
}
\examples{
get_filters("wcvp")
get_filters("powo")
get_filters("ipni")

}
