% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_keywords}
\alias{get_keywords}
\title{Get valid keywords for a resource.}
\usage{
get_keywords(resource = c("wcvp", "powo", "ipni", "tol"))
}
\arguments{
\item{resource}{The resource being queried.}
}
\value{
A character vector of valid keywords.
}
\description{
Get valid keywords for a resource.
}
\examples{
get_keywords("wcvp")
get_keywords("powo")
get_keywords("ipni")

}
