% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kew_citation.R
\name{kew_citation}
\alias{kew_citation}
\title{Get citation for Kew data resource.}
\usage{
kew_citation(x)
}
\arguments{
\item{x}{Result of a call to \code{\link[=search_powo]{search_powo()}}, \code{\link[=lookup_powo]{lookup_powo()}}, \code{\link[=search_wcvp]{search_wcvp()}},
\code{\link[=lookup_wcvp]{lookup_wcvp()}}, \code{\link[=search_ipni]{search_ipni()}}, \code{\link[=lookup_ipni]{lookup_ipni()}}, \code{\link[=search_tol]{search_tol()}}, \code{\link[=load_tol]{load_tol()}},
\code{\link[=match_knms]{match_knms()}}, or \code{\link[=match_krs]{match_krs()}}}
}
\value{
A citation object with a print method for nice display.
}
\description{
Given the result of a query to a Kew data resource, get the appropriate
citation.
}
\examples{
r <- search_powo(list(character="poison"))
kew_citation(r)

r <- lookup_powo("320035-2")
kew_citation(r)

r <- search_wcvp(list(genus="Poa"))
kew_citation(r)

r <- lookup_wcvp("320035-2")
kew_citation(r)

r <- search_ipni(list(publishing_author="L."))
kew_citation(r)

r <- lookup_ipni("12653-1")
kew_citation(r)

r <- search_tol("Poa")
kew_citation(r)

r <- lookup_tol(2717)
kew_citation(r)

tree <- load_tol()
kew_citation(tree)

match <- match_knms("Poa annua")
kew_citation(match)

match <- match_krs("Poa annua")
kew_citation(match)


}
