% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{load_tol}
\alias{load_tol}
\title{Load the Tree of Life or another file from ToL.}
\usage{
load_tol(url = NULL, .wait = 0.1)
}
\arguments{
\item{url}{URL pointing to a file on the ToL SFTP server.}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\description{
Request a tree file for the whole ToL or an alignment,
sequence, or gene tree for a particular specimen or gene.
}
\details{
The \href{https://treeoflife.kew.org/}{Tree of Life} is a database
of specimens sequenced as part of Kew's efforts to build
a comprehensive evolutionary tree of life for flowering plants.

Newick tree, alignment, and sequence files are help on an SFTP server
for download. The URLs to access these are stored in entries for specimens
and genes in the ToL database. These can be accessed by either using \code{\link[=search_tol]{search_tol()}}
to get all specimens for a particular order, family, genus, or species or by
looking up a specific specimen or gene using \code{\link[=lookup_tol]{lookup_tol()}}. If no URL is specified,
this will load the ToL tree.
}
\examples{
 # load the ToL
 load_tol()

 # load a specimen fasta file
 specimen_info <- lookup_tol("1296")
 load_tol(specimen_info$fasta_file_url)

 # load a gene alignment file
 gene_info <- lookup_tol("51", type="gene")
 load_tol(gene_info$alignment_file_url)

 # load the gene tree
 load_tol(gene_info$tree_file_url)

}
\references{
Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D., Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L., Kersey P.J., Leitch I.J. & Forest F. 2021. A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life. Systematic Biology, 2021; syab035, https://doi.org/10.1093/sysbio/syab035
}
\seealso{
\itemize{
\item \code{\link[=lookup_tol]{lookup_tol()}} to lookup information about a sequenced specimen
using a valid ToL ID.
\item \code{\link[=search_tol]{search_tol()}} to search ToL using taxonomic info.
\item \code{\link[=download_tol]{download_tol()}} to save a file on the ToL SFTP server to file.
}

Other ToL functions: 
\code{\link{download_tol}()},
\code{\link{lookup_tol}()}
}
\concept{ToL functions}
