% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipni.R
\name{lookup_ipni}
\alias{lookup_ipni}
\title{Look up a name in IPNI.}
\usage{
lookup_ipni(id, type = c("taxon", "author", "publication"))
}
\arguments{
\item{id}{A string containing a valid IPNI ID.}

\item{type}{The type of record to look up. Either \code{taxon}, \code{author}, or \code{publication}.}
}
\value{
An \code{ipni_taxon} object, which is a simple structure with fields
for each of the fields returned by the lookup API, as well as the the \link[httr:response]{httr response object}.
}
\description{
Request the record for a taxon, author, or publication name in IPNI,
using the IPNI ID.
}
\details{
The \href{https://www.ipni.org/}{International Plant Names Index (IPNI)}
is a service that provides nomenclatural information for vascular plant names.

The name lookup API allows users to retrieve information for
a specific taxon, author, or publication name using the unique IPNI ID. If this is not known,
it can be found out using the \link[=search_ipni]{IPNI search API}.
}
\examples{

# retrieve nomenclatural information for a taxon name
name <- lookup_ipni("271445-2", "taxon")
print(name)

# format the results in a table
format(name)

# retrieve nomenclatural information for an author
name <- lookup_ipni("20028192-1", type="author")
format(name)

# retrieve nomenclatural information for a publication
name <- lookup_ipni("987-2", type="publication")
format(name)

}
\seealso{
\itemize{
\item \code{\link[=search_ipni]{search_ipni()}} to search IPNI using a taxon name.
}

Other IPNI functions: 
\code{\link{search_ipni}()}
}
\concept{IPNI functions}
