% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{lookup_tol}
\alias{lookup_tol}
\title{Look up a sequenced specimen in ToL.}
\usage{
lookup_tol(specimenid, .wait = 0.1)
}
\arguments{
\item{specimenid}{A string containing a valid ToL specimen ID.}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\value{
A \code{tol_specimen} object, which is a simple structure with fields
for each of the fields returned by the lookup API,
as well as the the \link[httr:response]{httr response object}.
}
\description{
Request the record for a sequenced specimen in ToL using
its ToL ID.
}
\details{
The \href{https://treeoflife.kew.org/}{Tree of Life} is a database
of specimens sequenced as part of Kew's efforts to build
a comprehensive evolutionary tree of life for flowering plants.

The specimen lookup API allows users to retrieve taxonomic and sequencing
information for a specific sequenced specimen using the unique ToL ID.
If this is not known, it can be found out using the \link[=search_tol]{ToL search API}.
}
\examples{

# retrieve information for a particular specimen
lookup_tol("1296")

# print a summary of the returned information
r <- lookup_tol("1296")
print(r)

# tidy into a tibble
r <- lookup_tol("1296")
tidy(r)

# tidy the returned list of synonyms into a tibble
r <- lookup_tol("1296")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=synonyms, names_sep="_")

# expand the child entries returned for each entry
r <- lookup_wcvp("30000055-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=children, names_sep="_")

}
\references{
Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D., Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L., Kersey P.J., Leitch I.J. & Forest F. 2021. A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life. Systematic Biology, 2021; syab035, https://doi.org/10.1093/sysbio/syab035
}
\seealso{
\itemize{
\item \code{\link[=search_tol]{search_tol()}} to search ToL using taxonomic information.
}
}
\concept{ToL functions}
