% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knms.R
\name{match_knms}
\alias{match_knms}
\title{Match names with KNMS.}
\usage{
match_knms(names)
}
\arguments{
\item{names}{A list or character vector of taxon names for matching.}
}
\value{
A \code{knms_match} object - a simple structure containing the match
results and some statistics about the number of matches.
}
\description{
Use the Kew Names Matching Service to match taxon names to
records in Plants of the World Online.
}
\details{
The \href{http://namematch.science.kew.org/}{Kew Names Matching Service (KNMS)} allows
a user to submit taxon names for matching against records in
\href{http://powo.science.kew.org/}{Plants of the World Online (POWO)}.
As far as I can tell, it uses exact matching as well as some rules-based matching
to account for common orthographic variants and Latin mistakes.

Names can be submitted to KNMS with or without an author string.
If a name can match to multiple different records, for instance
with synonyms, KNMS will return multiple matches. As such, we recommend
submitting names first with the taxonomic authority and then without
if no match can be found.

KNMS allows multiple names to be submitted at once. However, it can
be slow in returning results if too many names are submitted. For lots of names,
\href{http://namematch.science.kew.org/csv}{the website provides an interface for submitting a CSV file}.
}
\examples{

# match a name
match_knms("Poa annua L.")

# match a vector of names
names <- c("Myrcia guianensis", "Calyptranthes ranulphii", "Poa annua")
match_knms(names)

# format match results into a table
names <- c("Myrcia guianensis", "Bad plant", "Poa annua")
matches <- match_knms(names)
format(matches)

}
