% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krs.R
\name{match_krs}
\alias{match_krs}
\title{Match a name using KRS.}
\usage{
match_krs(query, .wait = 0.2)
}
\arguments{
\item{query}{The name to match using the reconciliation service.}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\value{
Returns an object of class \code{krs_match} that is a simple
structure with slots for:
\itemize{
\item \code{results}: the query results parsed into a list.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Use the Kew Reconciliation Service to match a name against IPNI.
}
\details{
The \href{http://data1.kew.org/reconciliation/about/IpniName}{Kew Reconciliation Service (KRS)}
allows name matching against IPNI using an Open Refine reconcilliation API.
}
\examples{
# Match a single name.
match_krs("Solanum sanchez-vegae")

}
\seealso{
\itemize{
\item \code{\link[=match_knms]{match_knms()}} to use simple matching for a vector of names.
}
}
