% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-next.R
\name{request_next}
\alias{request_next}
\alias{request_next.wcvp_search}
\alias{request_next.powo_search}
\alias{request_next.ipni_search}
\title{Request the next page of search results for a kewr results object.}
\usage{
request_next(object)

\method{request_next}{wcvp_search}(object)

\method{request_next}{powo_search}(object)

\method{request_next}{ipni_search}(object)
}
\arguments{
\item{object}{A kewr search results object.}
}
\value{
The next page of results as a kewr search object of
the same type.
}
\description{
Request the next page of search results for a kewr results object.
}
\section{Methods (by class)}{
\itemize{
\item \code{wcvp_search}: Request the next page of WCVP search results.

\item \code{powo_search}: Request the next page of POWO search results.

\item \code{ipni_search}: Request the next page of IPNI search results.
}}

\examples{
r <- search_wcvp("Poa")
request_next(r)


\donttest{
r <- search_powo("Poa")
request_next(r)
}

r <- search_ipni("Poa")
request_next(r)

}
