% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo.R
\name{search_powo}
\alias{search_powo}
\title{Search POWO.}
\usage{
search_powo(query, limit = 24)
}
\arguments{
\item{query}{The string to query POWO with.}

\item{limit}{The maximum number of records to return.}
}
\value{
Returns an object of class \code{powo_search} that is a simple
stucture with slots for:
\itemize{
\item \code{total}: the total number of results held in POWO for the query
\item \code{pages}: the total number of results pages for the query.
\item \code{limit}: the maximum number of results requested from the API, per page.
\item \code{cursor}: a cursor to retrieve the next page of results from the API.
\item \code{results}: the query results parsed into a list.
\item \code{query}: the query string submitted to the API.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Query Plants of the World Online for taxon information.
}
\details{
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}
is a database of information on the world's flora. It curates information from
published floras and other sources of floristic information.

The search API allows users to query the database using plant names,
geographic terms, and floristic characters. These can be queried using
keyword arguments. See arguments for all implemented keywords.

The API returns taxonomic information as well as species descriptions and
image locations if available. These results can be limited, for example to accepted species,
using filters. See arguments for all implemented filters.
}
