% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol.R
\name{search_tol}
\alias{search_tol}
\title{Search Kew's Tree of Life for specimens.}
\usage{
search_tol(query = "", limit = 50, page = 1, .wait = 0.2)
}
\arguments{
\item{query}{The string to query the database with.}

\item{limit}{An integer specifying the number of results
to return.}

\item{page}{An integer specify the page of results to request.}

\item{.wait}{Time to wait before making a requests, to help
rate limiting.}
}
\value{
Returns an object of class \code{tol_search} that is a simple
structure with slots for:
\itemize{
\item \code{total}: the total number of results held in ToL for the query.
\item \code{page}: the page of results requested.
\item \code{limit}: the maximum number of results requested from the API.
\item \code{results}: the query results parsed into a list.
\item \code{query}: the query string submitted to the API.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Query Kew's Tree of Life for specimens that have
been sampled for sequencing.
}
\details{
The \href{https://treeoflife.kew.org/}{Tree of Life} is a database
of specimens sequenced as part of Kew's efforts to build
a comprehensive evolutionary tree of life for flowering plants.

The search API allows users to query the database for specimens
based on their taxonomic information.
}
\examples{
# get the first 50 of all sequenced specimens
search_tol(limit=50)

# search for all sequenced Myrcia specimens
search_tol("Myrcia")

# get all sequenced specimens
search_tol(limit=5000)

# search for a species name and print the results
r <- search_tol("Myrcia guianensis")
print(r)

# simplify search results to a `tibble`
r <- search_tol("Myrcia")
tidy(r)

# gene stats are nested in the results
r <- search_tol("Myrcia")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=gene_stats)

# species names are nested in the results
r <- search_tol("Myrcia")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=species, names_sep="_")

# as is higher taxonomy
r <- search_tol("Myrcia")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=species, names_sep="_")

}
\references{
Baker W.J., Bailey P., Barber V., Barker A., Bellot S., Bishop D., Botigue L.R., Brewer G., Carruthers T., Clarkson J.J., Cook J., Cowan R.S., Dodsworth S., Epitawalage N., Francoso E., Gallego B., Johnson M., Kim J.T., Leempoel K., Maurin O., McGinnie C., Pokorny L., Roy S., Stone M., Toledo E., Wickett N.J., Zuntini A.R., Eiserhardt W.L., Kersey P.J., Leitch I.J. & Forest F. 2021. A Comprehensive Phylogenomic Platform for Exploring the Angiosperm Tree of Life. Systematic Biology, 2021; syab035, https://doi.org/10.1093/sysbio/syab035
}
\concept{ToL functions}
