% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{search_wcvp}
\alias{search_wcvp}
\title{Search WCVP for a taxon.}
\usage{
search_wcvp(query, filters = NULL, cursor = "*", limit = 50, .wait = 0.1)
}
\arguments{
\item{query}{The taxon string to search WCVP for. If using keywords,
the query must be formatted as a list.}

\item{filters}{Filter to apply to search results.
Multiple filters must be supplied as a character vector.}

\item{cursor}{A cursor returned by a previous search.
If used, the query and filter must be exactly the same.}

\item{limit}{An integer specifying the maximum number of results
to return.}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\value{
Returns an object of class \code{wcvp_search} that is a simple
structure with slots for:
\itemize{
\item \code{total}: the total number of results held in WCVP for the query
\item \code{cursor}: a cursor to retrieve the next page of results from the API.
\item \code{limit}: the maximum number of results requested from the API.
\item \code{results}: the query results parsed into a list.
\item \code{query}: the query string submitted to the API.
\item \code{filter}: the filter strings submitted to the API.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Query the World Checklist of Vascular Plants search API
for a taxon string.
}
\details{
The \href{https://wcvp.science.kew.org/}{World Checklist of Vascular Plants (WCVP)}
is a global consensus view of all known vascular plant species.
It has been compiled by staff at RBG Kew in consultation with plant
group experts.

The search API allows users to query the checklist for plant names.
Currently, it does not support partial or fuzzy matching.
In order to get a result, the user must supply a valid name string.
For example, 'Myrcia' and 'Myrcia guianensis' will return results,
but 'M' or 'Myr' will not.

There is some support for querying using keyword arguments. The API is
not currently documented, so only keywords that are definitely there have
been implemented. Use the \code{get_keywords} function to view a list of all implemented keywords.

The API will return taxonomic information (the family, authority, status, and rank)
of all names matching the query. These results can be limited, for example to accepted species,
using filters. Use the \code{get_filters} function to view a list of all implemented filters.
}
\examples{
# search for all entries containing a genus name
search_wcvp("Myrcia")

# search for all accepted species within a genus
search_wcvp("Myrcia", filters=c("species", "accepted"))

# search for up to 10,000 species in a genus
search_wcvp("Poa", filters=c("species"), limit=10000)

# search for all names in a family
search_wcvp(list(family="Myrtaceae"))

# search for genera within a family
search_wcvp(list(family="Myrtaceae"), filters=c("genera"))

# search for all names with a specific epithet
search_wcvp(list(species="guianensis"))

# search for a species name and print the results
r <- search_wcvp("Myrcia guianensis", filters=c("species"))
print(r)

# simplify search results to a `tibble`
r <- search_wcvp("Poa", filters=c("species"))
tidy(r)

# accepted name info is nested inside the records for synonyms
# simplify accepted name info to the name ID
r <- search_wcvp("Poa", filters=c("species"))
tidied <- tidy(r)
tidyr::unnest(tidied, cols=synonymOf, names_sep="_")

}
\references{
WCVP (2020). World Checklist of Vascular Plants, version 2.0. Facilitated by the Royal Botanic Gardens, Kew. Published on the Internet; http://wcvp.science.kew.org/
}
\seealso{
\itemize{
\item \code{\link[=lookup_wcvp]{lookup_wcvp()}} to lookup information about a taxon name
using a valid IPNI ID.
\item \code{\link[=download_wcvp]{download_wcvp()}} to download the entire WCVP.
}

Other WCVP functions: 
\code{\link{download_wcvp}()},
\code{\link{lookup_wcvp}()}
}
\concept{WCVP functions}
