% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo.R
\name{lookup_powo}
\alias{lookup_powo}
\title{Look up a taxon in POWO.}
\usage{
lookup_powo(taxonid, distribution = FALSE, .wait = 0.2)
}
\arguments{
\item{taxonid}{A string containing a valid IPNI ID.}

\item{distribution}{Include distribution in results (default \code{FALSE}).}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\value{
A \code{powo_taxon} object, which is a simple structure with fields
for each of the fields returned by the lookup API, as well as the the \link[httr:response]{httr response object}.
}
\description{
Request the record for a taxon in Plants of the World Online (POWO)
using the IPNI ID.
}
\details{
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}
is a database of information on the world's flora. It curates information from
published floras and other sources of floristic information.

The taxon lookup API allows users to retrieve information about
a specific taxon name using the unique IPNI ID. If this is not known,
it can be found out using the \link[=search_powo]{POWO search API}.
}
\examples{
\donttest{
# retrieve information for a taxon name
lookup_powo("271445-2")

# print a summary of the returned information
r <- lookup_powo("271445-2")
print(r)

# tidy returned record into a tibble
r <- lookup_powo("271445-2")
tidy(r)

# tidy the returned list of synonyms into a tibble
r <- lookup_wcvp("60447743-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=synonyms, names_sep="_")

# tidy the returned list of children into a tibble
r <- lookup_wcvp("30000055-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=children, names_sep="_")
}

}
\seealso{
\itemize{
\item \code{\link[=search_powo]{search_powo()}} to search POWO using a taxon name.
}

Other POWO functions: 
\code{\link{search_powo}()}
}
\concept{POWO functions}
