% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{lookup_wcvp}
\alias{lookup_wcvp}
\title{Look up a taxon in WCVP.}
\usage{
lookup_wcvp(taxonid, .wait = 0.1)
}
\arguments{
\item{taxonid}{A string containing a valid IPNI ID.}

\item{.wait}{Time to wait before making a request, to help
rate limiting.}
}
\value{
A \code{wcvp_taxon} object, which is a simple structure with fields
for each of the fields returned by the lookup API, as well as the the \link[httr:response]{httr response object}.
}
\description{
Request the record for a taxon in the World Checklist of
Vascular Plants (WCVP) using the IPNI ID.
}
\details{
The \href{https://wcvp.science.kew.org/}{World Checklist of Vascular Plants (WCVP)}
is a global consensus view of all known vascular plant species.
It has been compiled by staff at RBG Kew in consultation with plant
group experts.

The taxon lookup API allows users to retrieve taxonomic information for
a specific taxon name using the unique IPNI ID. If this is not known,
it can be found out using the \link[=search_wcvp]{WCVP search API}.
}
\examples{

# retrieve taxonomic information for a taxon name
lookup_wcvp("271445-2")

# print a summary of the returned information
r <- lookup_wcvp("271445-2")
print(r)

# tidy into a tibble
r <- lookup_wcvp("271445-2")
tidy(r)

# tidy the returned list of synonyms into a tibble
r <- lookup_wcvp("60447743-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=synonyms, names_sep="_")

# expand the child entries returned for each entry
r <- lookup_wcvp("30000055-2")
tidied <- tidy(r)
tidyr::unnest(tidied, cols=children, names_sep="_")

}
\references{
WCVP (2020). World Checklist of Vascular Plants, version 2.0. Facilitated by the Royal Botanic Gardens, Kew. Published on the Internet; http://wcvp.science.kew.org/
}
\seealso{
\itemize{
\item \code{\link[=search_wcvp]{search_wcvp()}} to search WCVP using a taxon name.
\item \code{\link[=download_wcvp]{download_wcvp()}} to download the entire WCVP.
}

Other WCVP functions: 
\code{\link{download_wcvp}()},
\code{\link{search_wcvp}()}
}
\concept{WCVP functions}
