% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipni.R
\name{search_ipni}
\alias{search_ipni}
\title{Search IPNI.}
\usage{
search_ipni(query, filters = NULL, limit = 50)
}
\arguments{
\item{query}{The string to query IPNI with.}

\item{filters}{Filter to apply to search results. Can be one
or more of \code{families}, \code{genera}, \code{species}, \code{infrafamilies},
\code{infragenera}, and \code{infraspecies}.}

\item{limit}{The maximum number of records to return.}
}
\value{
Returns an object of class \code{ipni_search} that is a simple
structure with slots for:
\itemize{
\item \code{total}: the total number of results held in POWO for the query
\item \code{pages}: the total number of results pages for the query.
\item \code{limit}: the maximum number of results requested from the API, per page.
\item \code{results}: the query results parsed into a list.
\item \code{query}: the query string submitted to the API.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Query the International Plant Names Index for nomenclatural information.
}
\details{
The \href{https://www.ipni.org/}{International Plant Names Index (IPNI)}
is a service that provides nomenclatural information for vascular plant names.
}
\examples{
# search for all names containing Poa annua
results <- search_ipni("Poa annua")

# format search results in a table
format(results)

# extract author team information for the search results
results_tbl <- format(results)
tidyr::unnest(results_tbl, cols=c(authorTeam), names_sep="_")

}
