% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo.R
\name{search_powo}
\alias{search_powo}
\title{Search POWO.}
\usage{
search_powo(query, filters = NULL, cursor = NULL, limit = 50)
}
\arguments{
\item{query}{The string to query POWO with. If using keywords,
the query must be formatted as a list and keywords must be one or more
of \code{family}, \code{genus}, \code{species}, \code{name}, \code{common_name}, \code{author}, \code{characteristic},
\code{summary}, \code{appearance}, \code{flower}, \code{fruit}, \code{leaf}, \code{inflorescence}, \code{seed},
\code{cloning}, \code{use}, \code{distribution.}}

\item{filters}{Filter to apply to search results. Can be one
or more of \code{accepted}, \code{families}, \code{genera}, \code{species},
\code{infraspecies}, \code{has_images}.
Multiple filters must be supplied as a character vector.}

\item{cursor}{A cursor returned by a previous search.
If used, the query and filter must be exactly the same.}

\item{limit}{The maximum number of records to return.}
}
\value{
Returns an object of class \code{powo_search} that is a simple
structure with slots for:
\itemize{
\item \code{total}: the total number of results held in POWO for the query
\item \code{pages}: the total number of results pages for the query.
\item \code{limit}: the maximum number of results requested from the API, per page.
\item \code{cursor}: a cursor to retrieve the next page of results from the API.
\item \code{results}: the query results parsed into a list.
\item \code{query}: the query string submitted to the API.
\item \code{response}: the \link[httr:response]{httr response object}.
}
}
\description{
Query Plants of the World Online for taxon information.
}
\details{
\href{http://www.plantsoftheworldonline.org/}{Plants of the World Online (POWO)}
is a database of information on the world's flora. It curates information from
published floras and other sources of floristic information.

The search API allows users to query the database using plant names,
geographic terms, and floristic characters. These can be queried using
keyword arguments. See arguments for all implemented keywords.

The API returns taxonomic information as well as species descriptions and
image locations if available. These results can be limited, for example to accepted species,
using filters. See arguments for all implemented filters.
}
\examples{
\donttest{
# search for all entries containing a genus name
search_powo("Myrcia")

# search for all accepted species within a genus
search_powo("Myrcia", filters=c("species", "accepted"))

# search for up to 100 species in a genus
search_powo("Poa", filters=c("species"), limit=100)

# search for all names in a family
search_powo(list(family="Myrtaceae"))

# search for all accepted species with blue flowers
search_powo(list(flower="blue"), filters=c("accepted", "species"))

# search for all accepted genera in Mexico
search_powo(list(distribution="Mexico"), filters=c("accepted", "genera"))

# search for a species name and print the results
r <- search_powo("Myrcia guianensis", filters=c("species"))
print(r)

# simplify search results to a `tibble`
r <- search_powo("Poa", filters=c("species"))
format(r)
}

}
\seealso{
\itemize{
\item \code{\link[=lookup_powo]{lookup_powo()}} to look up a taxon in POWO using the IPNI ID.
}

Other POWO functions: 
\code{\link{lookup_powo}()}
}
\concept{POWO functions}
