% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{search_wcvp}
\alias{search_wcvp}
\title{Search WCVP for a taxon.}
\usage{
search_wcvp(query, filters = NULL)
}
\arguments{
\item{query}{The taxon string to search WCVP for.}

\item{filters}{Filter to apply to search results. Can be one
or more of \code{accepted}, \code{generic}, \code{specific}, \code{intraspecific}.
Multiple filters must be supplied as a character vector.}
}
\description{
Query the World Checklist of Vascular Plants search API
for a taxon string.
}
